meta-ni: Added a squashfs-tools-static recipe
Added a squashfs tools static recipe that is based on the Open Embedded
squashfs tools recipe but builds static and stripped executables.
These tools will be used on the host side during the build process of
other external components. Since we cannot be certain what version of
glibc is present on various build machines a static version is preferred.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>
Acked-by: Rich Tollerton <rich.tollerton@ni.com>

Natinst-P4-Changelist-ID: 8514703
Natinst-ReviewBoard-ID: 50288