Export host side squashfs-tools for other components
Created a new utils targets for exporting host side utilities to be used
by other external components during the build process.

Currently the only utility exported is mksquashfs. Both 32 and 64 bit
versions are exported. Building the x86 32-bit version required adding a
new i386 machine .conf. The machine name i386 is slightly misleading
since it actually is tuned for i586 but compatible with i386, and i486.
This is an unfortunate quirk of our internal build process which uses
i386 as a generic name for x86 32-bit targets.

Natinst-P4-Changelist-ID: 8514784
Natinst-ReviewBoard-ID: 50288

Acked-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>
Acked-by: Rich Tollerton <rich.tollerton@ni.com>
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>