niconsole-image: make /home/root a symlink to /home/admin
Some x11/xfce packages appear to be erroneously hardcoding the home
directory to be /home/$USER, which is not valid when booting an NILRT
image without NIAuth: currently, the username as listed in passwd is
still root, but its home directory is /home/admin. As a workaround, add
a symlink.

Natinst-ReviewBoard-ID: 51970
Acked-by: Haris Okanovic <haris.okanovic@ni.com>

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>