niconsole-image: inherit core-image for all images
Many of the IMAGE_FEATURES settings we care about come from
core-image.bbclass, which we had not been including, and so were not
getting included into our images. This notably includes "x11-base", so
that the qemu xfce image was not including video drivers.

The fix is to use core-image instead, but do it in the .inc file,
since it's going to be used by everybody.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>