xserver-xfce-init: allow startxfce4 execution as arbitrary user
startxfce4 is hardcoded to execute as UID=0, and assumes that UID's name
is "root" and its HOME is /home/root. This precludes starting XFCE as a
user and it doesn't work if UID=0 isn't named root.

Instead of handhacking USER and HOME, use `su` to construct a proper
login shell, and configure the user via an XUSER variable (can be
overridden in /etc/default/xserver-xfce). `su` is provided by util-linux.

Natinst-Rally-ID: US26916
Acked-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Scot Salmon <scot.salmon@ni.com>
Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>