udev-extraconf: copy from openembedded-core
udev-extraconf is brought into images by default, and appears to be
almost entirely composed of relatively machine- or distro-specific
stuff that we want to largely rewrite or remove.

It's a waste of effort to blacklist it out of images only to add an
NI-specific package back in; similarly, modifying it in a .bbappend
will probably conclude with rewriting all of it.

So I'm just copying the whole thing, and bumping the PR.

Natinst-Rally-ID: TA44427
Natinst-ReviewBoard-ID: 58630
Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>