udev-extraconf: use a valid no-op program for wl12xx RUN
The following console boot message is observed on myRIO:

udevd[829]: failed to execute '/lib/udev/' '': Permission denied

This is because we set RUN="" for wl12xx, to bypass module load. This
gets the job done, but because it is a nonexistent file, udev reports
an error.

To fix, run /bin/true.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>