packagegroup-ni-extra: new group for repo packages
This is meant to be a list of all the packages we "know" should build,
and are worth posting on a repo. (As opposed to building world, whose
success is much less controllable.)

I basically compiled this by adding anything in oe-core and meta-oe that
looked interesting.

I commented out the packages which appear to require x11, for the
benefit of the zynq builds which don't include x11 in DISTRO_FEATURES.

Some packages I thought were interesting didn't build, I commented those
out with the tag "fail".

For presently unknown reasons, every package in
packagegroup-self-hosted-* requires libx11, even though
packagegroup-self-hosted-graphics is the only one that explicitly
depends on it. To work around this, I've included the contents of those
packagegroups' RDEPENDS here.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>