xfce: Fix build dependency on `niacctbase`
Make the dependency on `niacctbase` recipe non-conditional.
`xfce-nilrt-settings` relies on the existence of lvuser:ni for the do_install
step in order to set the folder permissions correctly. A non-conditional
dependency on `niacctbase` insures that the do_populate_sysroot step in
`niacctbase` happens before the do_configure step in xfce-nilrt-settings.
This fixes a build error related to the user:group not being present.
Additionally DEPENDS_${PN} is not used in any other recipe in OpenEmbedded.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
--
Natinst-ReviewBoard-ID: 59632