xfce: Disable the X screen saver
Disabling just the power management for the monitor via 'xset -dpms' it is not
sufficient to prevent the monitor from going blank when the Intel i915 driver
is used. We also need to disable the X screen saver via 'xset s off' command.
Add a new autostart entry to do that.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Brad Mouring <brad.mouring@ni.com>
--
Natinst-ReviewBoard-ID: 60684