udev: detect sdcard insert/delete and mount device(s)
The sdcard interface on smasher essentially looks like a usb card
reader, so the fact that a USB device is present is understood by
udev, however no events for ejecting the sdcard are generated.

The solution is to turn on kernel polling for the device and detect
the "change" event to allow the various devices to be detected and
mounted by udev.

Natinst-CAR-ID: 431208
Natinst-Rally-ID: TA51103
Natinst-ReviewBoard-ID: 61213

Signed-off-by: Bill Pittman <bill.pittman@ni.com>
Acked-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Aaron Rossetto <aaron.rossetto@ni.com>