niwatchdogpet: Create a watchdog petter
Currently, the only way to pet the NI watchdog is to call an IOCTL.
Calling IOCTLs from shell scripts is prohibitevely difficult. We create
this simple petting tool to pet the watchdog. This tool will be used by
an init script to pet the watchdog during the boot process. We want to
pet the watchdog during the boot process in order to give the boot more
time to complete before the watchdog causes LEDs to blink indicating
that the boot is stuck because it is taking too long.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>
Natinst-CAR-ID: 475537