niconsole-image-safe: Include niwatchdogpet
Only include the niwatchdogpet init script in safemode, in order to
isolate changes to the minimum possible. We don't believe it would hurt
to have this in the ni-base packagegroup, but since we are so close to
our release date, we are taking the conservative approach and only
putting it in where we have to.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>
Natinst-CAR-ID: 475537