florence: Remove ability to close keyboard
Do not allow the user to exit Florence. If the user closes Florence
and has no access to a physical keyboard, it may be possible that
they cannot create a new instance of Florence. It is expected that
the user would either have no need for a virtual keyboard or would
know how to access Florence if a physical keyboard is attached.

Acked-by: Ben Shelton <ben.shelton@ni.com>
Signed-off-by: Jonathan David <jonathan.david@ni.com>
---
Natinst-ReviewBoard-ID: 72680