initscripts: Fix issue where /etc/init.d/urandom is overwritten
meta-nilrt provides a 1-line file to be sourced by the urandom init
script to set RANDOM_SEED_FILE. The file in meta-nilrt was originally
called "urandom", which is the same name as the init script itself in
openembedded-core, so it inadvertently overrides the init script with
the 1-liner, resulting in an image that includes a urandom "init script"
that just sets RANDOM_SEED_FILE. This of course results in the intended
actions of the urandom init script not being performed, including not
saving off the random seed between boots.

Fix it by moving meta-nilrt's urandom 1-liner to urandom.default and
installing it to the correct location.

Signed-off-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>
Acked-by: Alejandro del Castillo <alejandro.del.castillo@ni.com>
Acked-by: Scot Salmon <scot.salmon@ni.com>
---
Natinst-CAR-ID: 484436
Natinst-ReviewBoard-ID: 74941