xfce: Enable touch screen calibration on first run
Remove the .desktop file that was used to disable the auto-run for
xinput_calibrator. For 13.0 (Dylan) we had to disable it due to errors when
run as a non-root user. In the latest OE branches (e.g. Daisy) the
xinput_calibrator_once.sh script is fixed and can be run now as a non-root
user. The script will run the xinput_calibrator program if a touch screen is
detected and no calibration data is found or if the calibration data
corresponds to a different screen model.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>
---
Natinst-ReviewBoard-ID: 77985