xfce: Run xinput_calibrator wrapper script in a terminal
Change the auto-run .desktop file to run the xinput_calibrator wrapper shell
script in a terminal. The script is run on first run if the currently
connected touch screen is not calibrated. If the script is not run from a
terminal the X start process doesn't wait for the calibration to finish before
launching the rest of the Xfce windowing system and the calibration screen
disappears due to the window/desktop manager starting on top of it.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>
---
Natinst-ReviewBoard-ID: 79948