nisystemformat: manage possible error
nisystemformat can get an error from fuser, seemingly a race condition
where the process using the mount point dies between the time fuser
checks for running processes on the mount point and when it sends them
the kill signal. While it's odd and annoying that a process which was
told to die earlier took this long to actually die, regardless, passing
the error about failing to kill anything out of nisystemformat is not
helpful and just causes unnecessary failure of the entire script, so
eat the error. (Note that "|| :" is a higher-performance version of
"|| true" for this purpose.)

Signed-off-by: Scot Salmon <scot.salmon@ni.com>
Acked-by: Rich Tollerton <rich.tollerton@ni.com>
Natinst-CAR-ID: 495523