gcc: make default tunings match TUNE_CCARGS
Previously the default ARM options to gcc were being set to default values.
This didn't matter for building within OE since the proper options were set
on the command line, but when gcc is used on the target without supplying these
options the default options were producing non-optimal binaries.

Fix this problem by changing the default config options in gcc to match the
TUNE_CCARGS set for the machine.

Signed-off-by: Ken Sharp <ken.sharp@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Rich Tollerton <rich.tollerton@ni.com>
---
Natinst-ReviewBoard-ID: 86946