packagegroup-ni-runmode: Build CP936 for x64 and ARM
Undoing commit: packagegroup-ni-runmode: Only build CP936 for x64
This is a remnant from when we were using the Code Sourcery toolchain.
When we were using the third party toolchain, we were locked into a
specific version of glibc and had to have a specific binary format for
the CP936 codepage shared library which was checked into p4 and
installed on the target via CDF.  Now that we are no longer using the
third party toolchain, we can allow OE to build the codepage files and
install them in the image.

Signed-off-by: Bill Pittman <bill.pittman@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>
---
Natinst-ReviewBoard-ID: 88581