coreutils: add rdepends on broken out packages
We break out hostname, ls, and chcon from the coreutils package.
Currently, when installing the coreutils package, it does not install
these utilities, but it does redirect the update-alternatives symlinks
to point to them.  This causes 'ls' not to work after installing the
coreutils package alone.

Add an RDEPENDS on the broken-out packages to get back the behavior
users expect: installing the coreutils package installs all the
coreutils.

Signed-off-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>
---
Natinst-ReviewBoard-ID: 88585