packagegroup-ni-base: add niwatchdogpet
niwatchdogpet is currently included in safemode images only.  However,
it is needed in runmode to pet the watchdog once in order to give
LabVIEW an extra 30 seconds to start.  This may be necessary in the case
where all available NI software is installed and the boot process takes
longer.

Add it to packagegroup-ni-base so it is included in both safemode and
runmode images.

Signed-off-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Alejandro del Castillo <alejandro.del.castillo@ni.com>
Acked-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>
---
Natinst-CAR-ID: 514238
Natinst-ReviewBoard-ID: 90758