ni-utils: replace efimgr with efivar in fw_printenv
efimgr is a patched version of efiboomgr 0.6 and has a lot of hard-coded
information. The read and write variables part of efibootmgr evolved into a
different library called efivar and efibootmgr 0.11 depends on it. Since our
patch for efibootmgr is significantly large compared to the patch on efivar,
beacuse efibootmgr depends anyway on efivar and efivar is aleady pulled in OE,
efivar will replace efimgr for OS to EFI communication on NI Linux targets.

Signed-off-by: Alexandru But <alexandru.but@ni.com>
Acked-by: Adrian Papp <adrian.papp@ni.com>
---
Natinst-ReviewBoard-ID: 90395