busybox: defconfig: provide logger
In both the 2014 and 2015 releases, we provide logger in x64 runmode --
it's pulled in by one of the XFCE dependencies.  We should provide it
for Zynq as well, and also in safemode, so users can make entries in the
system log.

Provide this using BusyBox to minimize the footprint required.

Signed-off-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Will Page <will.page@ni.com>
---
Natinst-ReviewBoard-ID: 91892