kernel: Removing NI's kernel references
The NI kernel recipes reference our internal git server. Removing
the references for now as this impedes recipe parsing.

As a note, currently the image building process only uses the kernel
for providing packages to the feeds, namely the debug kernel and the
perf tool. The actual kernel used in the installable images is still
built separately.

Signed-off-by: Brad Mouring <brad@the-bradlands.net>