busybox: use pidfile properly
Currently, running '/etc/init.d/busybox-acpid restart' will print a
message that there was no running daemon to stop, and then spawn a
second copy of busybox-acpid.  The issue is that acpid is provided by
busybox on our targets, so it's run as /bin/busybox.nosuid, not
/sbin/acpid, which doesn't match.

To fix the issue, tell start-stop-daemon about the pidfile that acpid
already creates, and specify process name rather than executable path.
Additionally, use the '--retry' option when stopping the daemon so it
will wait until the daemon exits before returning -- this prevents
a race condition in the restart case where start-stop-daemon sees the
terminated-but-still-running acpid as 'process already running' and does
not start the daemon again.

Signed-off-by: Ben Shelton <ben.shelton@ni.com>
Reviewed-by: Ken Sharp <ken.sharp@ni.com>
Reviewed-by: Rich Tollerton <rich.tollerton@ni.com>
---
Natinst-ReviewBoard-ID: 95342