openssh: Remove code that conditionally starts sshd based on .ini settings
Traditionally we have enabled/disabled sshd via a token in ni-rt.ini with the
default being 'disabled'. However the new way of provisioning targets via salt
requires sshd to be enabled and running by default. Remove the code that
checks for the [SystemSettings]/sshd.enabled token and exits early if not set.

A yet to be defined method of disabling sshd after provisioning will be provided
once the UI for it on the host side becomes available. Note that sshd can still
be disabled on target via the standard way: i.e by creating a file named
/etc/ssh/sshd_not_to_be_run.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Natinst-ReviewBoard-ID: 98290