coreutils: Fix circular dependency build error
coreutils depends on niacctbase which in turn depends indirectly on
coreutils (through useradd.bbclass). Move the offending instructions to
pkg_postinst so they run after the users/groups are added, but keep the
hostname rename in the install_append or the pkg coreutils-hostname will
not be created (exe not found).

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>