linux-nilrt: Add recipe for kernel built with full dynamic ticks
Add recipe for building a kernel with full dynamic ticks support (nohz).

Additionally remove redundant package group for the debug kernel - it was
initially intended to be extensible to other kernels but due to bitbake dynamics
when building a kernel (i.e. requiring a 'clean' step between builds) it is
superfluous now.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Natinst-ReviewBoard-ID: 103709