Revert "packagegroup-ni-mono-extra: add mono-xsp"
This reverts commit 0f2eb0e31a87d0cdb9316ff3fd163576d2ef0476.

mono-xsp 3.0.11 does not build with mono4 because it requires gmcs,
which is not installed by mono. (This was masked on my dev machine
because I had symlinked /usr/bin/gmcs to /usr/bin/mcs). There are fixes
committed in upstream git, but until then, we should disable building
this to avoid breaking the build.