xilinx-zynq.conf: Fix ${MACHINE_ARCH} value
In openembedded-core/meta/conf/bitbake.conf there is this line:

MACHINE_ARCH = "${@[d.getVar('TUNE_PKGARCH', True), d.getVar('MACHINE',
 True)][bool(d.getVar('MACHINE', True))].replace('-', '_')}"

It replaces any dash with an underscore in ${MACHINE_ARCH} and this is
bad for the xilinx-zynq build which expects a dash in the variable.
Therefore we overwrite this replace ONLY for xilinx-zynq machines.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>