nilrt.conf: set IMAGE_LINGUAS for en-us
Currently, images are built with the default IMAGE_LINGUAS?="en-us
en-gb" setting in image.bbclass. We have no need for en-gb, but we would
like en-us.ISO-8859-1, in order to facilitate spot checking execution of
LANG=L1, etc. So override this in distro.conf so that it applies for any
image that doesn't set IMAGE_LINGUAS itself.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>
Natinst-Reviewboard-ID: 109197
Natinst-CAR-ID: 507802
Reviewed-by: Ken Sharp <ken.sharp@ni.com>