nisetlocale.sh: remove nirtcfg; use locale.conf; set LANG=en_US
Use the de-facto standard of /etc/locale.conf to store system-wide
locale settings.

Default to LANG=en_US instead of LANG=POSIX, because (a) that's the only
locale we expect to ship with at present, and (b) we lack
infrastructure (i.e. separate package or runtime code) to install
/etc/locale.conf, so we might as well default LANG to what we actually
want. (In the future, when L10N is properly supported, this will
probably change back to POSIX.)

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>
Natinst-CAR-ID: 507802
Natinst-Reviewboard-ID: 109604
Reviewed-by: Ken Sharp <ken.sharp@ni.com>