initscripts: Filter characters used in generated hostname
If a product includes 'Embedded Controller' in the name, strip
that from the default hostname. Also, VMWare images return serial#
values with spaces and with a length greater than necessary for
our initial autogenerated value, so limit the characters used to
the last eight usable values for consistency with serial numbers
of NI controllers.

Signed-off-by: Joshua Hernstrom <joshua.hernstrom@ni.com>
Natinst-Reviewboard-ID: 110320