xserver-xfce-init: remove support for disabling via nirtcfg
We are working towards deprecating the use of ni-rt.ini. Controlling X
via ni-rt.ini is not implemented for the next release and will happen
through a currently-unknown implementation mechanism for future
releases. Anybody who needs to do this in the short term will be able to
accomplish the same thing by disabling the initscript directly,
uninstalling packages, etc.

Natinst-Reviewboard-ID: 112910
Reviewed-by: Ken Sharp <ken.sharp@ni.com>
Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>