nilrt.conf: Set prefered provider for virtual/fftw to fftwf
fftw and fftwf are in conflict; they are identical only fftwf enables
support for neon float operations. gnuradio depends on fftwf, and we
explicitly require fftw in packagegroup-ni-extra where gnuradio is also
added. Thus a build error occurs when building both at the same time.

I see no reason to use fftw instead of fftwf, so switch to it.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>