crio-support-scripts: rewrite nisetfpgaautoload to avoid nirtcfg
We are working towards deprecating the use of ni-rt.ini. The
current mechanism for configuring this setting (autoload of FPGA user
bitfile) is getting ripped out, which makes the ni-rt.ini token
configuring this behavior dead.

We don't anticipate needing this setting for the next release, but
simply keying it off of the existence of a /etc/nofpga file is as good
of a solution as any. So rewrite this initscript to do that.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>
Natinst-Reviewboard-ID: 112913
Reviewed-by: Scot Salmon <scot.salmon@ni.com>