nisetcommitratio: rewrite to use one file per reservation
We are working towards deprecating the use of ni-rt.ini. Memory
reservation accounting currently uses INI tokens stored in ni-rt.ini,
and is a much-used feature; however, it is not required for the next
release.

Instead of requiring installers to manipulate INI files in order to
update memory reservations, it would be more ideal to let each installer
manage their own file containing their own reservation, much like how
/etc/cron.d is better for packages than /etc/crontab. We can implement
this now, while nobody is relying on the behavior, in order to reduce
dependencies on ni-rt.ini.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>
Natinst-Reviewboard-ID: 112911
Reviewed-by: Scot Salmon <scot.salmon@ni.com>