sysconfig-settings{,-ui}: move to extras
The existing sysconfig-settings facility is explicitly getting rewritten
and we have been requested to break it now to figure out what's broken.
Remove it from the runmode packagegroup (thus keeping it out of the
images), but keep it in the extras packagegroup, in case somebody needs
to install it later. (We will delete the recipes at a later time.)

Natinst-Reviewboard-ID: 113145
Reviewed-by: Ken Sharp <ken.sharp@ni.com>
Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>