ni-utils: add ninetcfgutil
ninetcfgutil is a utility for saving, restoring, and resetting the
network configuration for NI's current network configuration stack. It
is most notably used to manage the network config across formats and
system imaging, both of which are in the process of getting rewritten
such that this utility will not be needed.

However, it is also used for IP Reset, which is *not* going away at the
moment; so we still need ninetcfgutil for at least a little while longer.

This is mostly an import from Perforce, with some path changes to
install to /usr/bin and /usr/lib instead of /usr/local/natinst/bin and
/etc/natinst/networking.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>