host-toolchain-sysroot: Clean-up dangling and escaping symlinks
Clean-up dangling and escaping symlinks from SDK sysroots which cannot
be dealt with in recipes:

ld.so.cache symlink escapes sysroot. It's generated by one or more
packages running ldconfig during sysroot creation.

mtab and lock point into unmounted file systems. These would be
mounted during boot on a system, but not in SDKs.

lsb installs symlinks pointing to an optional package that's not
installed in SDKs.

Sysroot with dangling or escaping symlinks can't be packaged with
`tar -h`, which is useful for building Windows SDKs.

Signed-off-by: Haris Okanovic <haris.okanovic@ni.com>
Natinst-ReviewBoard-ID: 117604