xilinx-zynqhf.conf: specify UBOOT related parameters
Set UBOOT_ARCH to arm and specify a UBOOT_MACHINE.
The UBOOT_MACHINE is set to nicrio9068, because the
fw_printenv/setenv tools require a valid config
from the U-Boot tree to be built. Since the fw_printenv/
setenv tools are user-space utilities, they will be
compatible with all xilinx-zynqhf targets.
IMAGE_CLASSES holds the classes that the image class
inherits. To create a cpio.gz.u-boot type ramdisk we
need to add the image_types_uboot to the IMAGE_CLASSES
variable. The restore-mode ramdisk will be created
as a cpio.gz.u-boot ramdisk.

Signed-off-by: Arnold Csorvasi <arnold.csorvasi@ni.com>