rtctl: move init script from late rcS.d to early rc5.d
The first boot init script was running a little too early and needed to be
moved later in the initialization.  For comms 2.0, just move it to rc5.d.
Specifically, rcS.d/S99rtctl was running before rcS.d/S99run-postinst due to
the whole alphabetical ordering thing and leaving the target in a state where
it really needed to be rebooted again so rtctl could catch the new interrupts
installed by drivers' post-installs.

Signed-off-by: Bill Pittman <bill.pittman@ni.com>