nilrt.conf: set LOCALE_UTF8_IS_DEFAULT=0 and adjust IMAGE_LINGUAS
Presently on Yocto, the locale en_US is of UTF-8 encoding; the
ISO-8859-1 locale is en_US.ISO-8859-1. This is backwards from glibc and
most/all other Linux distributions, where "en_US" is ISO-8859-1 and the
UTF-8 locale is "en_US.UTF-8".

This is not normally a problem, except that Python has been observed to
hardcode the assumption that en_US always means ISO-8859-1. The
following output remains the same regardless of the actual encoding of
the en_US locale:

$ LANG=en_US python -c 'import locale; print(locale.getdefaultlocale());'
('en_US', 'ISO8859-1')

To work around this, and maximize compatibility with desktop Linux,
set LOCALE_UTF8_IS_DEFAULT=0, and adjust IMAGE_LINGUAS for the renamed
locales.

Note that at present, because the default locale is en_US, this commit
will change the default encoding to ISO-8859-1. A separate commit will
change the default back to UTF-8.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>
Natinst-CAR-ID: 507802
Natinst-ReviewBoard-ID: 124364
Reviewed-by: Adrian Ratiu <adrian.ratiu@ni.com>