migrate-nilrt: add migration commands
This package is a wrapper around all migration images that will be used to
switch between LV and LVNext target setups. After installation, the user will be
able to call "ni_migrate_target" and the script will start the migration
process depending on the present state.
Each command will detect the right migration image to pull for the target and
than execute the task to switch paradigm.

Signed-off-by: Alexandru But <alexandru.but@ni.com>