zynq-bootscripts: create bootscripts containing booting logic
Create 4 script files that contain the booting logic. The gen_bs.sh
script will create 4 text files, which will be converted to U-Boot
script files with mkimage.
The boot.scr file is the top level bootscript. It checks for the
requested bootmode and loads the appropriate default/safemode/restore.scr
file. The default.scr contains the commands to boot into the default
image. The safemode.scr is separated from the default one to handle
extra requirements for booting into the safemode. The restore.scr
knows only about the booting logic for restore mode and lives on the
bootfs partition.