zynq-itb: create runmode, safemode and migrate fake itb
Create linux_safemode.itb and linux_runmode.itb for
redirecting U-Boot to /boot/boot.scr. linux_safemode.itb
is required for safemode booting via the power button.
The names are kept for compatibility reasons with the
existing U-Boot environment. Both itbs try to redirect U-Boot
to /boot/boot.scr which contain the booting logic, and in case
it is missing, try the restore.scr from the bootfs partition.
If both are missing, stop at the U-Boot console.
Create a linux_migrate.itb, which will be installed on the target
as linux_runmode.itb, and contains the isntructions to boot the
migration image installed by opkg.