apache2/mod_ssl: Fixup protocol discovery hooks on Windows/MSVC
This is a fixup to change 79d4d4d3 that was backported from master as
npn-patch-2.4.7.patch. It's not broken on Linux because of the way
DSO imports/exports are declared, but Windows/MSVC's dllimport/dllexport
is less permissive.

This change may be dropped whenever afa9bc99 makes it into a release.

More details in 0002-mod_ssl-Fixup-protocol-discovery-hooks-on-Windows-MS.patch

Signed-off-by: Haris Okanovic <haris.okanovic@ni.com>
Natinst-Rally-ID: US84115
Natinst-CAR-ID: 519365
Natinst-ReviewBoard-ID: 128648
Natinst-ReviewBoard-ID: 128326