apr: enable pthread support (moved from oe-core)
Apache2 needs pthread support (pthread_kill especially) during its
configure step to fix a signal hang at runtime. Pthread support needs
to be added to apr in order for the webserver to correctly find it.

This was initially done using autotools cached configurations in OEcore.
Upstream refused it because they need apr_cv_process_shared_works=yes,
so this is a downstream NI-specific configuration.

Instead of setting directly the arch caches in OE-core, move this to the
apr recipe's CACHED_CONFIGUREVARS inside a bbappend in meta-nilrt.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>