curl: fix for update to 7.44.0
The vxworks config file was slightly modified in the update, change
the patch file to apply cleanly with these changes. In addition,
move away from using a wildcard for the curl .bbappend file.

Remove patch #0020, commit was cherry-picked from a later version
of cURL to fix an issue. That fix is now part of the source code
in 7.44.0.

Signed-off-by: Jonathan David <jonathan.david@ni.com>