zynq-itb: create linux-bw-migrate.itb
This itb is used to trigger backwards migration on arm devices which
boot an itb directly as opposed to using bootscripts and having the
kernel, device tree, rootfs on a filesystem (as in the newer nilrt OS).
The itb contains just the minimal bootscript necessary to trigger the
migration process.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>