mk-NI-hybridiso.wks: Fix restore-mode-image on vm
Append restore=provision to kernel parameters to generated isolinux boot
configuration.

Previously when attempting to boot off an iso on a virual machine the vm
would fail to provision due to a missing restore parameter in the iso
bootloader. This adds the missing parameter to the iso bootloader
allowing it to properly provision on virtual machines.

Signed-off-by: Collin Richards <collin.richards@ni.com>
Natinst-CAR-ID: 590707
Natinst-ReviewBoard-ID: 141927