ni_provisioning: don't restore arm cfg partition
Provisioning wipes the contents of this partition but there's no need to
actually restore it from a pre-stored zip file because on boot the arm
safemode cfgs populate it if it's empty.

Thus this step is redundant: we might as well remove it to simplify the
recipe and scripts by removing safemode_default_config.tar.gz altogether.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>