gtk+/gtk3+.inc: remove docbook-utils-native dep
docbook-utils [1] was opensourced by redhat in 2011 and then it was
abandoned. Since it is unmaintaned for so many years is breaks on
newer distros in a lot of ways. The only package using it in OE is
gnome-doc-utils (which we don't use btw) but it is added as a
dependency for the entire gtk+ stack. This is an error which should
be corrected in OE upstream. Even gnome-doc-tools builds/runs without
this dependency, so it is dead weight and should be removed.

We do not build it right now and do not depend on gtk+.inc in any
way but we have stuff like the florence keyboard which is disabled
but should still be buildable. Disabling docbook-utils-native in the
main gtk+ recipe allows us to build gtk+ dependent recipes.

When OE upstream decides to remove docbook-utils the line in our
bbappends will have no effect and should also be removed.

[1] https://git.fedorahosted.org/cgit/docbook-utils.git/

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>