mkfs.ext: force write ext partition filesystems
mke2fs is the tool we use to create ext filesystems during provision.
After v1.43 (included in krogoth and beyond) it adds an interactive
user prompt if a partition we intend to write already exists on disk.
(commit dbcd617882ee5 in the e2fsprogs-code repo).

This is very annoying to our use case because during provisioning we
prompt the user at the start of the process whether we wants to continue
and mke2fs stops the provisioning process to ask the user again.

There are two ways to override this default behaviour: set a timeout to
default to the [y]es answer or force the write. Setting the timeout to
1 second (smallest amount) is undesirable because we introduce a time
interval when the user can accidentaly press 'n' and leave the target
in a semi-provisioned state. So we always force rewrite the filesystem.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>