gnome-icon-theme: remove recipe
We are carrying this recipe in meta-nilrt even though we don't use it
for nothing. We had a bbappend for it before it was removed from OE
entirely so I moved it to meta-nilrt thinking it might be useful. This
seems to have been used way back in 2013/2014, but not since then.

So many things are wrong with it that it's better to also remove it
and use the adawita fork (which OE upstream replaced it with) when we
want to put out an xfce NG image.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>