linux-nilrt: update to 4.6
These recipes are not used to build the kernel which we ship with
older nilrt distros; that one is built by a build tool named NIBuild.
OE needs some kind of kernel recipes to satisfy its dependency for
virtual/kernel and this was the sole purpouse of these old 3.14
recipes besides random debugging/development.

However, with the update to gcc 5.3 contained in OE krogoth, kernel
3.14 does not build anymore so we remove it and upgrade to a more
recent 4.6 branch. Also remove the nohz version since we don't provide
such a branch.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>