nilrt.conf: remove qmmp from QTDEMOS due to licensing
meta-qt4 ships some demo applications which are included in packagegroup-qt-core
which in turn gets included in packagegroup-ni-extra. qmmp is one of those apps
but besides the gplv2 source it also has a commercial clause asking for a
"commercial" flag to be whitelisted in the distro.

Since we're not using qmmp at anything and it is better to avoid licensing crap,
le't just remove the package from the default demos variable.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>