nilrt.conf: explicitly set GCCVERSION
In this config it is not explicitly set like in the CG nilrt-xfce.conf or
NG's nilrt.conf. We want to use GCC 5.3 and because it's the highest version
in these branches it is selected implicitely, but if we are to upgrade and
want to keep 5.3 this would screw us because OE would silently update gcc.

So set the GCCVERSION explicitely to avoid any headaches.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>