files/group */passwd: add dhcp group & user
The dhcp recipe starting from krogoth adds its own user & group to
allow starting as "dhcp:dhcp" via a config option which is disabled
by default. Even if the config is disabled the user is added at the
image build time and since we're using static id's for this branch
it will fail. Fix this by adding dhcp:dhcp.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>