init-restore-mode: hack to disable intel cstates
Because we're experiencing severe performance degradations when provisioning
cRIO 903[012] because cstates are enabled by default in BIOS we explicitely
disable them in restoremode. This is already done in NG runmode via the
nidisablecstates initrd script and this is why we haven't seen this there.

However, restoremode does not have an init system, our restore script is our
PID 1 so we have to add this hack here.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>