restore-mode-image: fix race with minimal-nilrt-image
We're including minimal-nilrt-image in restore-mode-image but we're not
depending on it and this causes the build to sometimes fail if restore
tries to search for the minimal image in the sysroot but the minimal
image build hasn't finished.

By adding minimal-nilrt-image to RDEPENDS we ensure that it is built and
present in the sysroot (the do_rootfs[depends] dependency is not enough).

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>