nilrt*conf: add VIRTUAL-RUNTIME_xserver_common = xserver-common
The opkg update showed us that the x11-common and xserver-common ipk's
were overwriting eachother's files because OE-core set the runtime
provider to x11-common and we needed xserver-common in our meta-nilrt
xserver-xfce-init recipe.

This error was uncaught by the previous opkg versions, so set the
runtime provider also in our distro conf to the one we need to fix it.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>