curl: Backport CVE-2016-5419 fix from cURL 7.50.1
Bug: libcurl would attempt to resume a TLS session even if the client
certificate had changed. That is unacceptable since a server by
specification is allowed to skip the client certificate check on resume,
and may instead use the old identity which was established by the
previous certificate (or no certificate).

Backported fix: Switch off SSL session id when client cert is used.

https://curl.haxx.se/docs/adv_20160803A.html

Signed-off-by: Haris Okanovic <haris.okanovic@ni.com>
Natinst-CAR-ID: 599706
Natinst-ReviewBoard-ID: 157131